divided PathDecision
